Keep registry threads asleep until data arrives
Fixes EEF-70

When background monitoring is enabled, the Terminal registry thread
consumes all available CPU time. A similar problem in the CAN registry
thread was worked around in 30ea4cc.

In both registry threads, an Event is now used to signal when data
arrives or when the interface is closing and the thread should shut
down. Arguably, letting Queue block would be simpler, but requires an ad
hoc handshake to tell the registry threads to shut down.